Extract AddCloudLibrariesDialog to top level
